+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 
 - [2026-04-01 09:32] fix(core): fix memory leak by cancelling core subscription
+- [2026-04-01 09:34] feat(theme): implement responsive theme view component